feat: beerlotTitle updated
