+++ src/components/feed/FeedTemplate.tsx
@@ -1,14 +1,8 @@
-import {
-  Box,
-  Button,
-  ButtonProps,
-  Container,
-  useDisclosure,
-} from "@chakra-ui/react";
+import { Box, Container, useDisclosure } from "@chakra-ui/react";
 import Cookies from "js-cookie";
 import { useRouter } from "next/router";
-import { EditPencil } from "../../../public/svg";
 import { CenteredTitle } from "../shared/Headers/CenteredTitle";
+import { ReviewModalTriggerButton } from "../shared/ReviewModal/ReviewModalTriggerButton";
 import { ReviewModalWrapper } from "../shared/ReviewModal/ReviewModalWrapper";
 import { FeedTabList } from "./FeedTabList";
 
@@ -35,27 +29,5 @@ export const FeedTemplate = () => {
         <ReviewModalWrapper isModalOpen={isOpen} onCloseModal={onClose} />
       </Container>
     </Box>
-  );
-};
-
-interface ReviewModalTriggerButtonProps extends ButtonProps {}
-
-const ReviewModalTriggerButton: React.FC<ReviewModalTriggerButtonProps> = ({
-  ...props
-}) => {
-  return (
-    <Button
-      w="70px"
-      h="70px"
-      pos="fixed"
-      borderRadius="full"
-      bg="orange.300"
-      bottom={100}
-      right={"10vw"}
-      _hover={{}}
-      {...props}
-    >
-      <EditPencil />
-    </Button>
   );
 };
