+++ src/components/shared/Headers/LeftCloseRandom.tsx
@@ -1,7 +1,7 @@
-import {Box, Button, Flex, Text} from "@chakra-ui/react";
+import { Box, Button, Flex, Text } from "@chakra-ui/react";
 import React from "react";
-import {X} from "../../../../public/svg";
-import BeerlotTitle from "./BeerlotTitle";
+import { X } from "../../../../public/svg";
+import { BeerlotTitle } from "./BeerlotTitle";
 
 interface LeftCloseRandomProps {
   title?: string;
